# api-spec.yaml
openapi: 3.0.3
info:
  title: hits-classroom API
  version: 0.3.0
  description: Backend API для LMS (курсы, задания, материалы, посты, оценки, файлы).
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /users/{userId}/files:
    get:
      tags: [files]
      summary: Список загруженных файлов пользователя
      operationId: listUserFiles
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/File"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Files ─────────────────────────────────────────────────────────────────────
  /files:
    post:
      tags: [files]
      summary: Загрузка файла
      operationId: uploadFile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          description: Файл слишком большой
...
  /files/{fileId}/info:
    get:
      tags: [files]
      summary: Получение информации о файле (без самого файла)
      operationId: getFileInfo
      security:
        - bearerAuth: []
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Courses ───────────────────────────────────────────────────────────────────
...
  /courses/{courseId}/posts/{postId}/files/{fileId}:
    get:
      tags: [files]
      summary: Скачать файл, прикреплённый к посту. Доступно всем участникам курса.
      operationId: getPostFile
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CourseId"
        - $ref: "#/components/parameters/PostId"
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Содержимое файла
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Пост или файл не найдены, либо файл не прикреплён к данному посту
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /courses/{courseId}/materials/{materialId}/files/{fileId}:
    get:
      tags: [files]
      summary: Скачать файл, прикреплённый к материалу. Доступно всем участникам курса.
      operationId: getMaterialFile
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CourseId"
        - $ref: "#/components/parameters/MaterialId"
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Содержимое файла
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Материал или файл не найдены, либо файл не прикреплён к данному материалу
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /courses/{courseId}/assignments/{assignmentId}/submissions/{submissionId}/files/{fileId}:
    get:
      tags: [submissions]
      summary: >-
        Скачать файл, прикреплённый к решению.
        Owner/teacher видят файлы любого решения.
        Студент — только из своего решения.
      operationId: getSubmissionFile
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CourseId"
        - $ref: "#/components/parameters/AssignmentId"
        - name: submissionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Содержимое файла
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Решение или файл не найдены, либо файл не прикреплён к данному решению
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AssignmentId:
      name: assignmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PostId:
      name: postId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MaterialId:
      name: materialId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    File:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        file_name:
          type: string
        file_size:
          type: integer
          format: int64
        mime_type:
          type: string
        created_at:
          type: string
          format: date-time

    # ── Courses ───────────────────────────────────────────────────────────────
...